Moved execute and query logic to modules
https://github.com/CosmWasm/cw-plus/pull/804#discussion_r979888833
