+++ contracts/cw1155-base/src/contract.rs
@@ -1,27 +1,13 @@
-use crate::{
-    error::ContractError,
-    msg::InstantiateMsg,
-    state::{APPROVES, BALANCES, MINTER, TOKENS},
-};
+use crate::{error::ContractError, execute, msg::InstantiateMsg, query, state::MINTER};
 use cosmwasm_std::{
-    entry_point, to_binary, Addr, Binary, Deps, DepsMut, Env, MessageInfo, Order, Response,
-    StdResult, SubMsg, Uint128,
+    entry_point, to_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult,
 };
-use cw1155::{
-    ApproveAllEvent, ApprovedForAllResponse, BalanceResponse, BatchBalanceResponse,
-    Cw1155BatchReceiveMsg, Cw1155ExecuteMsg, Cw1155QueryMsg, Cw1155ReceiveMsg, Expiration,
-    IsApprovedForAllResponse, TokenId, TokenInfoResponse, TokensResponse, TransferEvent,
-};
+use cw1155::{Cw1155ExecuteMsg, Cw1155QueryMsg};
 use cw2::set_contract_version;
-use cw_storage_plus::Bound;
-use cw_utils::{maybe_addr, Event};
 
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:cw1155-base";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-
-const DEFAULT_LIMIT: u32 = 10;
-const MAX_LIMIT: u32 = 30;
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
@@ -58,380 +44,51 @@ pub fn execute(
             token_id,
             value,
             msg,
-        } => execute_send_from(env, from, to, token_id, value, msg),
+        } => execute::send_from(env, from, to, token_id, value, msg),
         Cw1155ExecuteMsg::BatchSendFrom {
             from,
             to,
             batch,
             msg,
-        } => execute_batch_send_from(env, from, to, batch, msg),
+        } => execute::batch_send_from(env, from, to, batch, msg),
         Cw1155ExecuteMsg::Mint {
             to,
             token_id,
             value,
             msg,
-        } => execute_mint(env, to, token_id, value, msg),
-        Cw1155ExecuteMsg::BatchMint { to, batch, msg } => execute_batch_mint(env, to, batch, msg),
+        } => execute::mint(env, to, token_id, value, msg),
+        Cw1155ExecuteMsg::BatchMint { to, batch, msg } => execute::batch_mint(env, to, batch, msg),
         Cw1155ExecuteMsg::Burn {
             from,
             token_id,
             value,
-        } => execute_burn(env, from, token_id, value),
-        Cw1155ExecuteMsg::BatchBurn { from, batch } => execute_batch_burn(env, from, batch),
+        } => execute::burn(env, from, token_id, value),
+        Cw1155ExecuteMsg::BatchBurn { from, batch } => execute::batch_burn(env, from, batch),
         Cw1155ExecuteMsg::ApproveAll { operator, expires } => {
-            execute_approve_all(env, operator, expires)
+            execute::approve_all(env, operator, expires)
         }
-        Cw1155ExecuteMsg::RevokeAll { operator } => execute_revoke_all(env, operator),
+        Cw1155ExecuteMsg::RevokeAll { operator } => execute::revoke_all(env, operator),
     }
-}
-
-/// When from is None: mint new coins
-/// When to is None: burn coins
-/// When both are None: no token balance is changed, pointless but valid
-///
-/// Make sure permissions are checked before calling this.
-fn execute_transfer_inner<'a>(
-    deps: &'a mut DepsMut,
-    from: Option<&'a Addr>,
-    to: Option<&'a Addr>,
-    token_id: &'a str,
-    amount: Uint128,
-) -> Result<TransferEvent<'a>, ContractError> {
-    if let Some(from_addr) = from {
-        BALANCES.update(
-            deps.storage,
-            (from_addr, token_id),
-            |balance: Option<Uint128>| -> StdResult<_> {
-                Ok(balance.unwrap_or_default().checked_sub(amount)?)
-            },
-        )?;
-    }
-
-    if let Some(to_addr) = to {
-        BALANCES.update(
-            deps.storage,
-            (to_addr, token_id),
-            |balance: Option<Uint128>| -> StdResult<_> {
-                Ok(balance.unwrap_or_default().checked_add(amount)?)
-            },
-        )?;
-    }
-
-    Ok(TransferEvent {
-        from: from.map(|x| x.as_ref()),
-        to: to.map(|x| x.as_ref()),
-        token_id,
-        amount,
-    })
-}
-
-/// returns true iff the sender can execute approve or reject on the contract
-fn check_can_approve(deps: Deps, env: &Env, owner: &Addr, operator: &Addr) -> StdResult<bool> {
-    // owner can approve
-    if owner == operator {
-        return Ok(true);
-    }
-    // operator can approve
-    let op = APPROVES.may_load(deps.storage, (owner, operator))?;
-    Ok(match op {
-        Some(ex) => !ex.is_expired(&env.block),
-        None => false,
-    })
-}
-
-fn guard_can_approve(
-    deps: Deps,
-    env: &Env,
-    owner: &Addr,
-    operator: &Addr,
-) -> Result<(), ContractError> {
-    if !check_can_approve(deps, env, owner, operator)? {
-        Err(ContractError::Unauthorized {})
-    } else {
-        Ok(())
-    }
-}
-
-pub fn execute_send_from(
-    env: ExecuteEnv,
-    from: String,
-    to: String,
-    token_id: TokenId,
-    amount: Uint128,
-    msg: Option<Binary>,
-) -> Result<Response, ContractError> {
-    let from_addr = env.deps.api.addr_validate(&from)?;
-    let to_addr = env.deps.api.addr_validate(&to)?;
-
-    let ExecuteEnv {
-        mut deps,
-        env,
-        info,
-    } = env;
-
-    guard_can_approve(deps.as_ref(), &env, &from_addr, &info.sender)?;
-
-    let mut rsp = Response::default();
-
-    let event = execute_transfer_inner(
-        &mut deps,
-        Some(&from_addr),
-        Some(&to_addr),
-        &token_id,
-        amount,
-    )?;
-    event.add_attributes(&mut rsp);
-
-    if let Some(msg) = msg {
-        rsp.messages = vec![SubMsg::new(
-            Cw1155ReceiveMsg {
-                operator: info.sender.to_string(),
-                from: Some(from),
-                amount,
-                token_id: token_id.clone(),
-                msg,
-            }
-            .into_cosmos_msg(to)?,
-        )]
-    }
-
-    Ok(rsp)
-}
-
-pub fn execute_mint(
-    env: ExecuteEnv,
-    to: String,
-    token_id: TokenId,
-    amount: Uint128,
-    msg: Option<Binary>,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv { mut deps, info, .. } = env;
-
-    let to_addr = deps.api.addr_validate(&to)?;
-
-    if info.sender != MINTER.load(deps.storage)? {
-        return Err(ContractError::Unauthorized {});
-    }
-
-    let mut rsp = Response::default();
-
-    let event = execute_transfer_inner(&mut deps, None, Some(&to_addr), &token_id, amount)?;
-    event.add_attributes(&mut rsp);
-
-    if let Some(msg) = msg {
-        rsp.messages = vec![SubMsg::new(
-            Cw1155ReceiveMsg {
-                operator: info.sender.to_string(),
-                from: None,
-                amount,
-                token_id: token_id.clone(),
-                msg,
-            }
-            .into_cosmos_msg(to)?,
-        )]
-    }
-
-    // insert if not exist
-    if !TOKENS.has(deps.storage, &token_id) {
-        // we must save some valid data here
-        TOKENS.save(deps.storage, &token_id, &String::new())?;
-    }
-
-    Ok(rsp)
-}
-
-pub fn execute_burn(
-    env: ExecuteEnv,
-    from: String,
-    token_id: TokenId,
-    amount: Uint128,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv {
-        mut deps,
-        info,
-        env,
-    } = env;
-
-    let from_addr = deps.api.addr_validate(&from)?;
-
-    // whoever can transfer these tokens can burn
-    guard_can_approve(deps.as_ref(), &env, &from_addr, &info.sender)?;
-
-    let mut rsp = Response::default();
-    let event = execute_transfer_inner(&mut deps, Some(&from_addr), None, &token_id, amount)?;
-    event.add_attributes(&mut rsp);
-    Ok(rsp)
-}
-
-pub fn execute_batch_send_from(
-    env: ExecuteEnv,
-    from: String,
-    to: String,
-    batch: Vec<(TokenId, Uint128)>,
-    msg: Option<Binary>,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv {
-        mut deps,
-        env,
-        info,
-    } = env;
-
-    let from_addr = deps.api.addr_validate(&from)?;
-    let to_addr = deps.api.addr_validate(&to)?;
-
-    guard_can_approve(deps.as_ref(), &env, &from_addr, &info.sender)?;
-
-    let mut rsp = Response::default();
-    for (token_id, amount) in batch.iter() {
-        let event = execute_transfer_inner(
-            &mut deps,
-            Some(&from_addr),
-            Some(&to_addr),
-            token_id,
-            *amount,
-        )?;
-        event.add_attributes(&mut rsp);
-    }
-
-    if let Some(msg) = msg {
-        rsp.messages = vec![SubMsg::new(
-            Cw1155BatchReceiveMsg {
-                operator: info.sender.to_string(),
-                from: Some(from),
-                batch,
-                msg,
-            }
-            .into_cosmos_msg(to)?,
-        )]
-    };
-
-    Ok(rsp)
-}
-
-pub fn execute_batch_mint(
-    env: ExecuteEnv,
-    to: String,
-    batch: Vec<(TokenId, Uint128)>,
-    msg: Option<Binary>,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv { mut deps, info, .. } = env;
-    if info.sender != MINTER.load(deps.storage)? {
-        return Err(ContractError::Unauthorized {});
-    }
-
-    let to_addr = deps.api.addr_validate(&to)?;
-
-    let mut rsp = Response::default();
-
-    for (token_id, amount) in batch.iter() {
-        let event = execute_transfer_inner(&mut deps, None, Some(&to_addr), token_id, *amount)?;
-        event.add_attributes(&mut rsp);
-
-        // insert if not exist
-        if !TOKENS.has(deps.storage, token_id) {
-            // we must save some valid data here
-            TOKENS.save(deps.storage, token_id, &String::new())?;
-        }
-    }
-
-    if let Some(msg) = msg {
-        rsp.messages = vec![SubMsg::new(
-            Cw1155BatchReceiveMsg {
-                operator: info.sender.to_string(),
-                from: None,
-                batch,
-                msg,
-            }
-            .into_cosmos_msg(to)?,
-        )]
-    };
-
-    Ok(rsp)
-}
-
-pub fn execute_batch_burn(
-    env: ExecuteEnv,
-    from: String,
-    batch: Vec<(TokenId, Uint128)>,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv {
-        mut deps,
-        info,
-        env,
-    } = env;
-
-    let from_addr = deps.api.addr_validate(&from)?;
-
-    guard_can_approve(deps.as_ref(), &env, &from_addr, &info.sender)?;
-
-    let mut rsp = Response::default();
-    for (token_id, amount) in batch.into_iter() {
-        let event = execute_transfer_inner(&mut deps, Some(&from_addr), None, &token_id, amount)?;
-        event.add_attributes(&mut rsp);
-    }
-    Ok(rsp)
-}
-
-pub fn execute_approve_all(
-    env: ExecuteEnv,
-    operator: String,
-    expires: Option<Expiration>,
-) -> Result<Response, ContractError> {
-    let ExecuteEnv { deps, info, env } = env;
-
-    // reject expired data as invalid
-    let expires = expires.unwrap_or_default();
-    if expires.is_expired(&env.block) {
-        return Err(ContractError::Expired {});
-    }
-
-    // set the operator for us
-    let operator_addr = deps.api.addr_validate(&operator)?;
-    APPROVES.save(deps.storage, (&info.sender, &operator_addr), &expires)?;
-
-    let mut rsp = Response::default();
-    ApproveAllEvent {
-        sender: info.sender.as_ref(),
-        operator: &operator,
-        approved: true,
-    }
-    .add_attributes(&mut rsp);
-    Ok(rsp)
-}
-
-pub fn execute_revoke_all(env: ExecuteEnv, operator: String) -> Result<Response, ContractError> {
-    let ExecuteEnv { deps, info, .. } = env;
-    let operator_addr = deps.api.addr_validate(&operator)?;
-    APPROVES.remove(deps.storage, (&info.sender, &operator_addr));
-
-    let mut rsp = Response::default();
-    ApproveAllEvent {
-        sender: info.sender.as_ref(),
-        operator: &operator,
-        approved: false,
-    }
-    .add_attributes(&mut rsp);
-    Ok(rsp)
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn query(deps: Deps, env: Env, msg: Cw1155QueryMsg) -> StdResult<Binary> {
     match msg {
         Cw1155QueryMsg::Balance { owner, token_id } => {
-            to_binary(&query_balance(deps, owner, token_id)?)
+            to_binary(&query::balance(deps, owner, token_id)?)
         }
         Cw1155QueryMsg::BatchBalance { owner, token_ids } => {
-            to_binary(&query_batch_balance(deps, owner, token_ids)?)
+            to_binary(&query::batch_balance(deps, owner, token_ids)?)
         }
         Cw1155QueryMsg::IsApprovedForAll { owner, operator } => {
-            to_binary(&query_is_approved_for_all(deps, env, owner, operator)?)
+            to_binary(&query::is_approved_for_all(deps, env, owner, operator)?)
         }
         Cw1155QueryMsg::ApprovedForAll {
             owner,
             include_expired,
             start_after,
             limit,
-        } => to_binary(&query_approved_for_all(
+        } => to_binary(&query::approved_for_all(
             deps,
             env,
             owner,
@@ -439,129 +96,14 @@ pub fn query(deps: Deps, env: Env, msg: Cw1155QueryMsg) -> StdResult<Binary> {
             start_after,
             limit,
         )?),
-        Cw1155QueryMsg::TokenInfo { token_id } => to_binary(&query_token_info(deps, token_id)?),
+        Cw1155QueryMsg::TokenInfo { token_id } => to_binary(&query::token_info(deps, token_id)?),
         Cw1155QueryMsg::Tokens {
             owner,
             start_after,
             limit,
-        } => to_binary(&query_tokens(deps, owner, start_after, limit)?),
+        } => to_binary(&query::tokens(deps, owner, start_after, limit)?),
         Cw1155QueryMsg::AllTokens { start_after, limit } => {
-            to_binary(&query_all_tokens(deps, start_after, limit)?)
+            to_binary(&query::all_tokens(deps, start_after, limit)?)
         }
     }
-}
-
-pub fn query_balance(deps: Deps, owner: String, token_id: String) -> StdResult<BalanceResponse> {
-    let owner = deps.api.addr_validate(&owner)?;
-
-    let balance = BALANCES
-        .may_load(deps.storage, (&owner, &token_id))?
-        .unwrap_or_default();
-
-    Ok(BalanceResponse { balance })
-}
-
-pub fn query_batch_balance(
-    deps: Deps,
-    owner: String,
-    token_ids: Vec<String>,
-) -> StdResult<BatchBalanceResponse> {
-    let owner = deps.api.addr_validate(&owner)?;
-
-    let balances = token_ids
-        .into_iter()
-        .map(|token_id| -> StdResult<_> {
-            Ok(BALANCES
-                .may_load(deps.storage, (&owner, &token_id))?
-                .unwrap_or_default())
-        })
-        .collect::<StdResult<_>>()?;
-
-    Ok(BatchBalanceResponse { balances })
-}
-
-fn build_approval(item: StdResult<(Addr, Expiration)>) -> StdResult<cw1155::Approval> {
-    item.map(|(addr, expires)| cw1155::Approval {
-        spender: addr.into(),
-        expires,
-    })
-}
-
-pub fn query_approved_for_all(
-    deps: Deps,
-    env: Env,
-    owner: String,
-    include_expired: bool,
-    start_after: Option<String>,
-    limit: Option<u32>,
-) -> StdResult<ApprovedForAllResponse> {
-    let owner = deps.api.addr_validate(&owner)?;
-    let start_after = maybe_addr(deps.api, start_after)?;
-    let limit = limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT) as usize;
-    let start = start_after.as_ref().map(Bound::exclusive);
-
-    let operators = APPROVES
-        .prefix(&owner)
-        .range(deps.storage, start, None, Order::Ascending)
-        .filter(|r| include_expired || r.is_err() || !r.as_ref().unwrap().1.is_expired(&env.block))
-        .take(limit)
-        .map(build_approval)
-        .collect::<StdResult<_>>()?;
-
-    Ok(ApprovedForAllResponse { operators })
-}
-
-pub fn query_token_info(deps: Deps, token_id: String) -> StdResult<TokenInfoResponse> {
-    let url = TOKENS.load(deps.storage, &token_id)?;
-
-    Ok(TokenInfoResponse { url })
-}
-
-pub fn query_is_approved_for_all(
-    deps: Deps,
-    env: Env,
-    owner: String,
-    operator: String,
-) -> StdResult<IsApprovedForAllResponse> {
-    let owner_addr = deps.api.addr_validate(&owner)?;
-    let operator_addr = deps.api.addr_validate(&operator)?;
-
-    let approved = check_can_approve(deps, &env, &owner_addr, &operator_addr)?;
-
-    Ok(IsApprovedForAllResponse { approved })
-}
-
-pub fn query_tokens(
-    deps: Deps,
-    owner: String,
-    start_after: Option<String>,
-    limit: Option<u32>,
-) -> StdResult<TokensResponse> {
-    let owner = deps.api.addr_validate(&owner)?;
-    let limit = limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT) as usize;
-    let start = start_after.as_ref().map(|s| Bound::exclusive(s.as_str()));
-
-    let tokens = BALANCES
-        .prefix(&owner)
-        .keys(deps.storage, start, None, Order::Ascending)
-        .take(limit)
-        .collect::<StdResult<_>>()?;
-
-    Ok(TokensResponse { tokens })
-}
-
-pub fn query_all_tokens(
-    deps: Deps,
-    start_after: Option<String>,
-    limit: Option<u32>,
-) -> StdResult<TokensResponse> {
-    let limit = limit.unwrap_or(DEFAULT_LIMIT).min(MAX_LIMIT) as usize;
-    let start = start_after.as_ref().map(|s| Bound::exclusive(s.as_str()));
-
-    let tokens = TOKENS
-        .keys(deps.storage, start, None, Order::Ascending)
-        .take(limit)
-        .collect::<StdResult<_>>()?;
-
-    Ok(TokensResponse { tokens })
 }
