+++ contracts/cw1155-base/src/lib.rs
@@ -1,6 +1,9 @@
 pub mod contract;
 mod error;
+pub mod execute;
+pub mod helpers;
 pub mod msg;
+pub mod query;
 pub mod state;
 
 pub use crate::error::ContractError;
